refactor(config): make all clients share one global config
This is a first step towards enabling config hot-reloading.
